# sdk/typescript/openapi.snapshot.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Cmsify API",
    "description": "Headless CMS API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/picklists": {
      "get": {
        "tags": [
          "PickLists"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PickListSummaryResponsePagedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListSummaryResponsePagedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListSummaryResponsePagedResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PickLists"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PickListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PickListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/picklists/{id}": {
      "get": {
        "tags": [
          "PickLists"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PickLists"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PickListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PickListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PickLists"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/picklists/{id}/revisions/{revisionId}": {
      "get": {
        "tags": [
          "PickLists"
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "revisionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PickListOptionRequest": {
        "required": [
          "label",
          "value"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PickListOptionResponse": {
        "required": [
          "id",
          "label",
          "order",
          "value"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PickListRequest": {
        "required": [
          "name",
          "options",
          "slug"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickListOptionRequest"
            }
          }
        },
        "additionalProperties": false
      },
      "PickListResponse": {
        "required": [
          "currentVersionNumber",
          "id",
          "name",
          "options",
          "slug"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickListOptionResponse"
            }
          },
          "currentRevisionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "currentVersionNumber": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PickListSummaryResponse": {
        "required": [
          "currentVersionNumber",
          "id",
          "name",
          "optionCount",
          "slug"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "optionCount": {
            "type": "integer",
            "format": "int32"
          },
          "currentRevisionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "currentVersionNumber": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PickListSummaryResponsePagedResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickListSummaryResponse"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter a Cmsify user session token, API client token, or JWT bearer token.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}